Machine Tool Monitoring Platform v2.0/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
paho-mqtt==1.6.1